requirements.txt
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
filelock==3.3.2
grpcio==1.41.1
idna==3.3
joblib==1.1.0
jsonschema==4.2.0
lxml==4.6.4
msgpack==1.0.2
numpy==1.21.4
pandas==1.3.4
pandas-datareader==0.10.0
patsy==0.5.2
protobuf==3.19.1
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
ray==1.8.0
redis==3.5.3
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.1
six==1.16.0
sklearn==0.0
statsmodels==0.13.0
threadpoolctl==3.0.0
urllib3==1.26.7
textblob
scattertext
plotly
